Move MTL under User blueprint
