+++ src/modules/site-v2/templates/data/trait.html
@@ -366,7 +366,7 @@
                 type: "POST",
 
                 success: function(response) {
-                    window.location.href = "{{ url_for('data.my_trait_library') }}";
+                    window.location.href = "{{ url_for('user.my_trait_library') }}";
                 },
 
                 error: function(error) {
