+++ src/modules/site-v2/templates/user/my-results.html
@@ -66,7 +66,7 @@
       <div class="card h-100 shadow text-center pt-5 mt-5 mx-auto cardBorderBottom" style="min-height:20rem;">
         <i class="bi bi-{{ bi_data_icon }} text-info" aria-hidden="true"></i>
         <div class="card-body">
-          <h3 class="h4 text-dark"><a class="stretched-link" href="{{ url_for('data.my_trait_library') }}">
+          <h3 class="h4 text-dark"><a class="stretched-link" href="{{ url_for('user.my_trait_library') }}">
             My Trait Library
           </a></h3>
         </div>
